Exec java instead of running sh in entrypoint
On a "docker stop" the Spring Boot application was not stopped gracefully.
Indeed the SIGTERM (on "docker stop") was sent to the "sh" process and
not the "java" process. This delay considerably the stop of the container
because it waits the grace period and terminate with a SIGKILL.
This is not an issue in a standalone Spring Boot application. But when
running a Spring Boot application in Docker container in a Spring Cloud
platform with Eureka (@EnableEurekaClient), when the docker container
is stopped, it does not unregister from the discovery service server.

see : https://docs.docker.com/engine/reference/commandline/stop/

Fixes gh-36